requirements.txt
# Core dependencies
google-generativeai>=0.3.0
PyGithub>=2.1.0
GitPython>=3.1.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
pytest-mock>=3.0.0
responses>=0.20.0
